+++ examples/image_editing_overlay/overlay_loading_indicator.tsx
@@ -1,0 +1,16 @@
+import * as React from "react";
+import { Box, LoadingIndicator } from "@canva/app-ui-kit";
+import styles from "./overlay_loading_indicator.css";
+
+export const OverlayLoadingIndicator = () => (
+  <Box
+    display="flex"
+    width="full"
+    height="full"
+    justifyContent="center"
+    alignItems="center"
+    className={styles.overlayLoadingIndicator}
+  >
+    <LoadingIndicator size="large" />
+  </Box>
+);